feat: migrate project management to Sanity CMS with updated API and schema
