+++ app/layout.tsx
@@ -2,6 +2,10 @@ import React from "react"
 import { Poppins } from "next/font/google"
 import type { Metadata } from "next"
 import "./globals.css"
+import MainNav from "@/components/main-nav"
+import { ThemeProvider } from "@/components/theme-provider"
+import { Toaster } from "@/components/ui/toaster"
+import { headers } from "next/headers"
 
 const poppins = Poppins({
   subsets: ["latin"],
@@ -14,11 +18,20 @@ export const metadata: Metadata = {
     "Portfolio of Kasinadh Sarma, a cybersecurity enthusiast with a B.Tech in Cyber/Computer Forensics and Counterterrorism from Parui University.",
 }
 
-/**
- * Root layout — intentionally minimal so /studio route doesn't inherit
- * portfolio-specific UI (nav, theme provider, etc.).
- * Portfolio-specific layout lives in app/(main)/layout.tsx.
- */
+async function LayoutWrapper({ children }: { children: React.ReactNode }) {
+  const headersList = await headers()
+  const pathname = headersList.get("x-pathname") ?? ""
+  const isStudio = pathname.startsWith("/studio")
+
+  return (
+    <ThemeProvider attribute="class" defaultTheme="system" enableSystem>
+      {!isStudio && <MainNav />}
+      {children}
+      {!isStudio && <Toaster />}
+    </ThemeProvider>
+  )
+}
+
 export default function RootLayout({
   children,
 }: {
@@ -43,9 +56,8 @@ export default function RootLayout({
         />
       </head>
       <body className={`min-h-screen bg-background antialiased ${poppins.className}`}>
-        {children}
+        <LayoutWrapper>{children}</LayoutWrapper>
       </body>
     </html>
   )
 }
-
