+++ proxy.ts
@@ -1,0 +1,13 @@
+// middleware.ts
+import { NextResponse } from "next/server"
+import type { NextRequest } from "next/server"
+
+export function proxy(request: NextRequest) {
+  const response = NextResponse.next()
+  response.headers.set("x-pathname", request.nextUrl.pathname)
+  return response
+}
+
+export const config = {
+  matcher: ["/((?!_next/static|_next/image|favicon.ico).*)"],
+}
